+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.3.2] - 2026-07-31
+## [0.3.2] - 2026-07-30
 
 >  It's IMPORTANT that you rebake any `UNCell` data (use the commandlet) as calculations and existing data needs to be rebuilt.